Refactor utils for server and client to be in one protocol.rs for easier acces and less confusion
