+++ src/server/mod.rs
@@ -23,14 +23,13 @@ use tracing::{
 
 use crate::db::errors::TransientError;
 use crate::metadata::RespValue;
+use crate::protocol::{parse_bulk_string, parse_integer};
 use crate::server::commands::{
     Command,
     ParsedResponse
 };
 use crate::server::utils::{
     check_argument,
-    parse_bulk_string,
-    parse_integer
 };
 use crate::DB;
